+++ sudoku_solver/sudoku_solver/UndoStep.cs
@@ -1,11 +1,11 @@
 ﻿public struct UndoStep
 {
     public int CellIdx;
-    public int RemovedBit;
+    public long RemovedBit;
 
     public const short ValueAssignmentFlag = 0;
 
-    public UndoStep(int cellIdx, int removedBit)
+    public UndoStep(int cellIdx, long removedBit)
     {
         CellIdx = cellIdx;
         RemovedBit = removedBit;
